Add initial ThinCurr python wrapper/interface (#57)
- Add python examples for ThinCurr
 - Add support for path prefix in plotting I/O to enable multiple mesh output
 - Add "proper" hash to matrix files as checksum
 - Move eigenvalue and frequency-response solvers to module for python and command line usage
 - Update mode conversion to use HDF5 format for generated model
 - Convert to Lanczos method for ThinCurr eigenvalues
 - Add python meshing functionality to create toroidally periodic grids
 - Add methods to compute grids from toroidal mode and REGCOIL files
 - Add method to compute current regularization operator
 - Improve XML support in ThinCurr

+++ src/utilities/generate_doc.py
@@ -195,5 +195,6 @@ if __name__ == '__main__':
         for img in imgs:
            shutil.copy(img, "docs/generated/images/{0}".format(os.path.basename(img)))
         # Remove temporary files
-        shutil.rmtree("{0}_files".format(base_path))
+        if os.path.isdir("{0}_files".format(base_path)):
+            shutil.rmtree("{0}_files".format(base_path))
         os.remove(base_path+".md")
